geoProject/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
asgiref==3.7.2
    # via django
build==1.0.3
    # via pip-tools
click==8.1.7
    # via pip-tools
django==5.0
    # via
    #   django-filter
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   geo-Project (pyproject.toml)
django-filter==23.5
    # via geo-Project (pyproject.toml)
djangorestframework==3.14.0
    # via
    #   djangorestframework-simplejwt
    #   geo-Project (pyproject.toml)
djangorestframework-simplejwt==5.3.1
    # via geo-Project (pyproject.toml)
editables==0.5
    # via hatchling
hatchling==1.21.0
    # via geo-Project (pyproject.toml)
markdown==3.5.1
    # via geo-Project (pyproject.toml)
packaging==23.2
    # via
    #   build
    #   hatchling
pathspec==0.12.1
    # via hatchling
pip-tools==7.3.0
    # via geo-Project (pyproject.toml)
pluggy==1.3.0
    # via hatchling
psycopg2-binary==2.9.9
    # via geo-Project (pyproject.toml)
pyjwt==2.8.0
    # via djangorestframework-simplejwt
pyproject-hooks==1.0.0
    # via build
pytz==2023.3.post1
    # via djangorestframework
sqlparse==0.4.4
    # via django
trove-classifiers==2023.11.29
    # via hatchling
wheel==0.42.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
